Add  platform check,`OfferPaymentMode`, and `presentedOfferingIdentifier` (#371)
## Motivation

[CF-1329](https://revenuecats.atlassian.net/browse/CF-1329)

- Updated to Android `6.1.1`
- New `Purchases.sharedInstance.store` to limit
`purchaseSubscriptionOption()` only for `PLAY_STORE`
- New `OfferPaymentMode` on `PricingPhase` and put in mapper
- Need to show `presentedOfferingIdentifier` in mapper and pass into
purchase methods

## Description

- Bumped native Android version to `6.1.1`
- Added `offerPaymentMode` to the `PricingPhase` mapper
- Check `Purchases.sharedInstance.store` in
`purchaseSubscriptionOption()` only for `PLAY_STORE`
- Added `presentedOfferingIdentifier` to `StoreProduct` and
`SubscriptionOption` mapper
- `purchaseProduct()` and `purchaseSubscriptionOption()` have a new
`presentedOfferingIdentifier: String?` parameter
- Fixed `purchaseSubscriptionOption()` tests that were actually called
`purchaseProduct()` 🤦‍♂️

### Reapplying the `presentedOfferingIdentifier`

In the native Android SDK, the `presentedOfferingIdentifier` is a
property of `StoreProduct` and `SubscriptionOption`. When the users
called `purchase()` with one of these objects, that offering identifier
is used without the developer having to do anything.

Because the way hybrids works, the `StoreProduct`s and
`SubscriptionOption`s instances that get fetched and are aren't up to
the hybrid level, aren't the same instances that are purchased going
back down to the native layer. Because of this, the
`presentedOfferingIdentifier` needs to get passed to the hybrid common
`purchaseProduct()` and `purchaseSubscriptionOption()` methods and
reapplied.



[CF-1329]:
https://revenuecats.atlassian.net/browse/CF-1329?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ