[BC5] iOS rename `productCategory` to `product type` and `productType` to `productSubtype` (#377)
## Motivation

iOS counterpart to #376 

Rename `productCategory` to `product type` and `productType` to
`productSubtype` for `StoreProduct`

## Description

These were _not_ being used by any of the hybrid models (in Flutter,
React Native, etc). Even though this is a slightly weird rename for iOS,
Android and the other hybrids were using "Product Type" and "Purchase
Type" so this will bring naming similar on hybrids.

The only kind of breaking change this _could_ have is if developers in
the Javascript/Typescript environment were looking at these undocumented
`productCategory` and `productType` keys but this will be called out in
the changelogs.