`CI`: changed Xcode 13 job to 13.4
A recent iOS change broke compilation with Xcode 13.2 (#409). There's no point fixing it because apps can't be deployed with Xcode 13 anymore.
We do however need to continue testing with Xcode 13 until https://github.com/CocoaPods/CocoaPods/issues/11558 is actually fixed (it's closed but it's not fixed), so that we know deployments will work.