Move typescript interfaces to PHC (#455)
This PR moves most of the typescript interfaces from the react native
SDK to PHC. The interfaces are mostly a copy-paste of those in
react-native-purchases. I haven't copied the `Purchases` interface yet.
Will do that separately since that will be a bit trickier considering
that in react native we are using static functions.

Counterpart react native PR:
https://github.com/RevenueCat/react-native-purchases/pull/684

##### TODO
- [x] Make sure they will be usable by both react native and capacitor.
This includes using them within the hybrids and making sure they are
exposed to consumers. Confirmed these interfaces are visible in the
react native PR
- [x] Make sure the documentation is generated correctly: While typedoc
separates the documentation into an "external" visibility, we can make
it show those by default, so it should still be ok.
- [x] Add typescript linter
- [x] Add typescript API tests (Added in #459)
- [ ] Move rest of interfaces to PHC (Missing purchases interface, will
do that one separately)
- [x] Publish to NPM