Re-organize android library (#603)
This is a pretty big change in terms of changes, but it's mostly moving
files around. Until now, the `android` folder was both a module and a
project. That's against how gradle projects are usually structured,
where you have a root project and one or more modules.

In this PR, we change this organization to use the proper one of having
a root project with multiple modules. Now, we have a project with 2
modules:
- api-tests
- hybridcommon

Some advantages of this:
- Api tests are now actually separated from the library module, so they
can more accurately reflect what API is really available to consumers.
- Now we can separate the revenuecat ui library much more easily by
creating a different module (Will be added in a followup PR)