Remove core-ktx dependency (#608)
This dependency was added by Android Studio's wizard and I didn't check
whether it was needed (it wasn't). It actually caused a breaking change,
since this dependency requires apps to target Android 34. For some
reason I didn't encounter this when I was testing in flutter using
`includeBuild` but was able to repro after publishing 8.1.3-beta.5. I
tested this again by publishing to maven local and can confirm the fix
now.