fix(PaywallProxy): check if root view controller is already presenting
In case our root view controller is already presenting another view controller,
go to the top of the "view controller food chain" and present the paywall there.

This will fix the issue, where it was not possible to present a paywall on top
of a modal presentation style view controller.

Signed-off-by: Martin Mazein <internet+github@amazeing.net>