[CI] Remove requirement for pod push to succeed to continue release process (#1030)
Pushing new pods has been very brittle lately. It's been a problem since
our current CI doesn't allow to continue the workflow after the job
fails. This makes it so the release process can continue even if pushing
the pods fail.

Some notes about this:
- We are still running the pod lint command for the tests before
releasing
- There is still a dependency of the `phcui` library to `phc`. This is
because the first one depends on the second one, so it doesn't make
sense to continue if the first one failed.