Create `purchases-js-hybrid-mappings` module with `CustomerInfo` model mappings (#1104)
This PR creates a new npm library to be used to hold the mappings from
the `purchases-js` SDK to what the hybrids expect, so we can share that
mapping logic accross all our hybrids. This follows a similar line to
what the android and iOS PHC libraries do.

In follow-up PRs I will be adding more mappings for other models and
more methods to the `PurchasesCommon` class.