Store `appUserId` in `localStorage` for web hybrid support (#1147)
There was a difference in behavior in Flutter web from Flutter mobile.
That is, in mobile, we store the user id in disk, so we reuse the same
even if the app is reconfigured normally, but we don't store it on web.
This adds some logic to store the user id in local storage on web
hybrids, so the behavior is consistent.