Make UI events listeners map events on main thread in Android (#1262)
We changed mapping of CustomerInfo and Offerings values to be performed
in a background thread in
https://github.com/RevenueCat/purchases-hybrid-common/pull/1169.
However, this can cause events to be delivered out of order and cause
some issues. Like the ones in
https://github.com/RevenueCat/react-native-purchases/issues/1224.

This reverts those changes done in the listener methods so events are
sent in order. We will work on a different improvement to make sure
these events can be processed in a background thread and be delivered
sequentially.