Add purchase method with options iOS (#1311)
iOS equivalent of #1291 

As we're adding new purchase functions with a "builder-like" pattern in
hybrids, having a function that accepts a free-form map of options and
includes the validation of the parameter becomes more useful. The
downside is not having compile-time safety... However, it will remove
logic from the hybrids which is a good thing.